Moving kind directory
